requirements.txt
# Core dependencies
streamlit==1.39.0
python-dotenv==1.0.0

# Document processing
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3

# Vector database and embeddings (CPU-only)
chromadb==0.5.18
sentence-transformers==3.3.1

# LLM support (optional - for advanced features)
langchain==0.3.7
langchain-community==0.3.7

# Utilities
tqdm==4.67.1
pyyaml==6.0.2
